Added proper git history comparison, to avoid rebuilding archives when not necessary

+++ GitmoSharp.Test/GitmoTest.cs
@@ -2,6 +2,7 @@
 using IO = System.IO;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using GitmoSharp;
+using System.Threading.Tasks;
 
 namespace GitmoSharp.Test
 {
@@ -11,6 +12,7 @@ namespace GitmoSharp.Test
         private string[] paths = { 
                                      "Test/NotInitialized", //0
                                       "Test/Initialized",   //1
+                                      "Test/InitializedWithCommit", //2
                                       "Test"
                                  };
         [TestInitialize]
@@ -29,7 +31,12 @@ namespace GitmoSharp.Test
         public void TestClean()
         {
             foreach (var path in paths) {
-                IO.Directory.Delete(path, true);
+                if (IO.Directory.Exists(path)) {
+                    try {
+                        IO.Directory.Delete(path, true);
+                    }
+                    catch (Exception ex) { }
+                }
             }
         }
 
@@ -49,6 +56,36 @@ namespace GitmoSharp.Test
             g.Zip(archiveID, relativePathInRepository, outPath);
 
             Assert.IsTrue(IO.File.Exists("Test/theid.zip"));
+        }
+
+        [TestMethod]
+        public void TestZipDirectory_WithCommit()
+        {
+            string repositoryPath = paths[2];
+            IO.Directory.CreateDirectory(IO.Path.Combine(repositoryPath, "somedir"));
+
+            IO.File.WriteAllText(IO.Path.Combine(repositoryPath, "somedir", "afile.txt"), DateTime.Now.ToString());
+
+            Gitmo g = new Gitmo(repositoryPath);
+            g.CommitChanges("a file commit");
+            
+            string archiveID = "theid";
+            string relativePathInRepository = "somedir"; // whole thing
+            string outPath = "Test/out";
+
+            bool didCreateZip = g.Zip(archiveID, relativePathInRepository, outPath);
+            Assert.IsTrue(didCreateZip, "first zip not made");
+
+            didCreateZip = g.Zip(archiveID, relativePathInRepository, outPath);
+            Assert.IsFalse(didCreateZip, "second zip attempt still made a zip");
+
+            IO.File.WriteAllText(IO.Path.Combine(repositoryPath, "somedir", "asecondfile.txt"), DateTime.Now.ToString());
+            Task.Delay(1000).Wait(); // required delay to make sure the second commit has a different timestamp;
+            g.CommitChanges("second commit");
+            didCreateZip = g.Zip(archiveID, relativePathInRepository, outPath);
+            Assert.IsTrue(didCreateZip, "did not create the zip after the second try");
+
+            Assert.IsTrue(IO.File.Exists("Test/out/theid.zip"));
         }
 
         [TestMethod]
